Fixed snmp_configurator usage, autoload splited
