+++ cloudshell/networking/juniper/autoload/mibs/JUNIPER-SMI.py
